Stream data from Feather
Notable change is using a single BLE characteristic to stream data
from the Feather to receiver.

Previously, one characteristic is used for each data and receivers
are expected to poll read these characteristics for data change. This
creates a lot of round trip messages to be made.

The change made uses a single characteristic containing a byte array.
The characteristic is also configured to allow receivers to subscribe
to notifications whenever the data has changed.

This reduces the number of round trip messages significantly and
hence power consumption. The receiver also does not have to poll for
data and can listen for notifications instead.